Football Scores: Feat: Connect the collection widget up to the database so it shows the latest data and doesn't crash Fix: Handle loss of internet connection better by checking for it before querying api and showing a toast if no connection is found Refactor: Removed API key for submission and included instructions in a Readme for evaluator Refactor: Added comments to point out where changes refer to the User Feedback
Alexandria:
Refactor: Remove Camera Preview as it's no longer in use
Fix: Issue where it seemed only 'To Kill a Mockinbird' would scan (was coincidence). Caused by the scanner scanning ISBN 13 numbers as ISBN 10 so the checksum was incorrect. Method added to correct this, which works out what it should be
Refactor: Remove scanner directory as all the calls to the code are made to me.dm7.barcodescanner.zbar
Refactor: Added comments to point out where changes refer to the User Feedback